+++ anon-service.sh
@@ -2,7 +2,7 @@
 
 # ####################################################################
 # anon-service.sh
-# version 2.0
+# version 2.1
 # 
 # Transparent proxy through Tor and optionally DNSCrypt with  
 # Anonymized-DNS feature enabled.
@@ -33,7 +33,6 @@ repo=/etc/apt/sources.list.d/tor.list
 dnscrel="2.1.4"
 ## If necessary, change the path according to your system
 export netman=/etc/NetworkManager/NetworkManager.conf
-export resolved=/etc/systemd/resolved.conf
 tor=/etc/tor/torrc
 unbound=/etc/unbound/unbound.conf
 
@@ -60,7 +59,7 @@ printf '%s\n' "   5.  Display status service"
 printf '%s\n' "   6.  Enable service to start automatically at boot"
 printf '%s\n' "   7.  Stop service without removing files and setting"
 printf '%s\n' "   8.  Exit removing service files and settings from system"
-printf '%s\n' "   9.  Edit torrc file"
+printf '%s\n' "   9.  Edit configuration files"
 printf '%s\n' "   10. Install this script"
 echo " ";
 echo -n "  Choose: ";
@@ -70,7 +69,8 @@ case "$task" in
 ## Detect if X runs
 if ! timeout 1s xset q &>/dev/null; then
 echo " ";
-echo "==> No X server detected. Try to use command line option instead." >&2
+echo "==> No X server detected. Try to use command-line option instead." >&2
+echo "";
 sleep 5
 exit 1
 fi
@@ -98,6 +98,7 @@ fi
 if hash wmctrl 2>/dev/null; then
 wmctrl -c :ACTIVE:
 else
+echo "";
 echo "==> Sorry! Your system is not ready to complete this action";
 echo "==> Please, check if you have installed the necessary files";
 sleep 7
@@ -106,6 +107,7 @@ return 1
 fi
 ;;
 5)
+clear
 checking_service
 sleep 7
 menu
@@ -137,9 +139,7 @@ menu
 return 1
 else
 echo "";
-xterm -T "Torrc" -e "leafpad $root/torrc" > /dev/null 2>&1 
-echo "==> Please restart the service to apply changes";
-sleep 7
+editor
 menu
 fi
 ;;
@@ -167,24 +167,30 @@ if [ -e "menu" ]; then
 menu
 return 1
 else 
+echo "";
 exit 1
 fi
 fi
+echo "   #######################################################";
+echo "   #                  ANON-SERVICE SETUP                 #";
+echo "   #######################################################";
 ### Checking for network connection
+echo "";
+echo "==> Checking for internet connection"
 rm conn.txt > /dev/null 2>&1
 ping -c1 opendns.com > conn.txt 2>&1
 if ( ! grep -q "icmp_seq=1" conn.txt ); then
 rm conn.txt > /dev/null 2>&1
-echo "==> Please check your network connection!";
+echo "==> Please connect to a network!";
 sleep 5
 if [ -e "menu" ]; then
 menu
 return 1
-else 
+else
+echo ""; 
 exit 1
 fi   
 fi
-clear
 echo "==> Checking dependencies and preparing the system"
 rm -rf $root > /dev/null 2>&1
 adduser -q --disabled-password --gecos "" $owner > /dev/null 2>&1
@@ -194,12 +200,11 @@ mkdir -p $root/temp
 chmod -R 777 $root/temp
 apt-get update > $root/temp/apt.log
 if [ ! -e "menu" ]; then
-apt-get install -y curl wget psmisc nano apt-transport-https unbound > /dev/null
+apt-get install -y curl wget psmisc nano apt-transport-https unbound ifupdown > /dev/null
 else
-apt-get install -y curl wget xterm psmisc wmctrl leafpad apt-transport-https unbound > /dev/null
+apt-get install -y curl wget xterm psmisc wmctrl leafpad apt-transport-https unbound ifupdown > /dev/null
 fi
 sleep 1
-clear
 if [ -e tor_option1 ]; then
 install_tor_project
 elif [ -e tor_option2 ]; then
@@ -214,6 +219,7 @@ if hash tor 2>/dev/null; then
 touch $root/installed
 else
 echo "==> Sorry! The script cannot recognize your Tor package";
+echo "";
 exit 1
 fi
 else
@@ -224,6 +230,7 @@ echo "      3.I already have tor installed";
 echo " "
 echo -n  " Choose: ";
 read -e choose
+echo "";
 case "$choose"
 in 1)
 install_tor_project
@@ -241,10 +248,12 @@ if hash tor 2>/dev/null; then
 touch $root/installed
 else
 echo "==> Sorry! The script cannot recognize your Tor package";
+sleep 3
 if [ -e "menu" ]; then
 menu
 return 1
 else 
+echo "";
 exit 1
 fi
 fi
@@ -256,6 +265,7 @@ if [ -e "menu" ]; then
 menu
 return 1
 else 
+echo "";
 exit 1
 fi
 esac
@@ -280,16 +290,15 @@ cd $root
 rm *.md > /dev/null 2>&1
 rm *.md* > /dev/null 2>&1
 echo "==> Downloading public DNS resolvers list";
+sleep 1
 curl -L -O https://download.dnscrypt.info/dnscrypt-resolvers/v3/public-resolvers.md > /dev/null 2>&1
 echo "==> Downloading anonymized DNS relays list";
+sleep 1
 curl -L -O https://download.dnscrypt.info/dnscrypt-resolvers/v3/relays.md > /dev/null 2>&1
-### Backup systemd-resolved
-if [ ! -s "$root/resolved.bak" ]; then
-cp $resolved $root/resolved.bak > /dev/null 2>&1
-fi
-### Backup NetworkManager.conf
 if [ ! -s "$netman.bak" ]; then
-cp $netman $netman.bak
+if [ -s "$netman" ]; then
+cp $netman $netman.bak > /dev/null 2>&1
+fi
 fi
 cd $(cat $root/cpath)
 }
@@ -310,15 +319,16 @@ echo $target > $root/temp/os.txt
 fi
 done
 os=$(cat $root/temp/os.txt | sed -e 's/^[ \t]*//')
-echo "";
 if [[ "$os" == " " ]]; then
 echo "==> Sorry! Apparently your OS hasn't candidate in Tor Project";
 echo "==> repo...Please, re-run the script and choose other options";
+echo "";
 exit 1
 fi
 if curl --head --silent --fail https://deb.torproject.org/torproject.org/dists/$os/main/binary-i386/ > /dev/null 2>&1;
 then
 echo "==> Enabling $os repository";
+sleep 1
 rm $repo > /dev/null 2>&1
 rm $repo* > /dev/null 2>&1
 touch $repo
@@ -337,7 +347,7 @@ wget -qO- https://deb.torproject.org/torproject.org/A3C4F0F979CAA22CDBA8F512EE8C
 sleep 1
 ### Fixing gnupg ownership 
 gpgconf --kill dirmngr
-chown -R $USER ~/.gnupg
+chown -R $USER ~/.gnupg > /dev/null 2>&1
 cd
 echo "==> Checking repository"; 
 apt-get update > $root/temp/apt.log 
@@ -349,6 +359,7 @@ then
    echo "==> Please, try to enter the correct codename for the debian or ubuntu";
    echo "==> repository compatible with your OS...In doubt, search on internet!";
    echo "==> Warning: if the repository is not correct, the script could crash!"; 
+   echo "";
    echo -n "    Please, enter the codename (for example: buster): ";
    read -e codename 
    rm $repo
@@ -379,9 +390,13 @@ if [ -e "menu" ]; then
 menu
 return 1
 else 
+echo "";
 exit 1
 fi
 fi
+echo "   #######################################################";
+echo "   #              ANON-SERVICE CONFIGURATION             #";
+echo "   #######################################################";
 ### Disable tor and unbound starting at boot time
 systemctl disable unbound > /dev/null 2>&1
 systemctl disable tor > /dev/null 2>&1
@@ -396,11 +411,11 @@ touch $root/stp-service
 echo "0" > $root/stp-service
 dnscryptconf
 elif [ -e "$(cat $root/cpath)/temp/menu" ] || [ -e "$(cat $root/cpath)/temp/configure" ]; then
-clear
+echo " ";
 echo "==> Which type of transparent proxy do you prefer to use?";
 echo "      1. Standard transparent proxy";
-echo "      2. Trasparent proxy with DNSCrypt and Anonymized-DNS feature";
-echo " "
+echo "      2. Trasparent proxy with DNSCrypt";
+echo " ";
 echo -n  " Choose: ";
 read -e choose
 case "$choose" in 
@@ -423,15 +438,21 @@ if [ -e "$(cat $root/cpath)/temp/menu" ]; then
 configure
 return 1
 else 
+echo "";
 exit 1
 fi
 esac
 else
+echo "";
 echo "==> Sorry! Something went wrong...Please, report this issue to the project";
+echo "";
 exit 1
 fi
+if [ -e "configure_option1" ]; then
+echo "";
+fi
 echo "==> Configuring Tor";
-sleep 2
+sleep 1
 ### Configuring Tor
 cp $tor $root/torrc
 echo "Log notice file $root/notices.log" >> $root/torrc
@@ -439,19 +460,17 @@ echo "VirtualAddrNetworkIPv4 10.192.0.0/10" >> $root/torrc
 echo "AutomapHostsOnResolve 1" >> $root/torrc
 echo "TransPort 9040 IsolateClientAddr IsolateClientProtocol IsolateDestAddr IsolateDestPort" >> $root/torrc
 echo "DNSPort 5353" >> $root/torrc
-### Disabling dnsmasq and configure Network-Manager and systemd-resolved
+### Disabling unwanted services and configure Network-Manager (if exists)
+echo "==> Configuring system";
+sleep 1
+if [ -s $netman ]; then
 rm $root/NetworkManager.conf.temp > /dev/null 2>&1
 cp $netman.bak $root/NetworkManager.conf.temp
 cd $root
 chown $USER:$USER NetworkManager.conf.temp
 sed -i 's/^dns=dnsmasq/#&/' NetworkManager.conf.temp
-sed '/\[main\]/a dns=default' NetworkManager.conf.temp > NetworkManager.conf
-if [[ -s "$root/resolved.bak" ]]; then
-cp $root/resolved.bak $root/resolved.conf.temp
-chown $USER:$USER resolved.conf.temp
-sed -i 's/^DNSStubListener=yes/#&/' resolved.conf.temp
-echo "DNS=127.0.0.1" >> resolved.conf.temp
-echo "DNSStubListener=no" >> resolved.conf.temp
+sed -i '/\[main\]/a dns=none' NetworkManager.conf.temp
+sed '/dns=none/a rc-manager=unmanaged' NetworkManager.conf.temp > NetworkManager.conf 
 fi
 rm $root/iptables_rules.sh > /dev/null 2>&1
 touch $root/iptables_rules.sh
@@ -466,6 +485,8 @@ echo "_user_uid=\"999\"" >> $root/iptables_rules.sh
 echo "_virt_addr=\"10.192.0.0/10\"" >> $root/iptables_rules.sh
 # Tor's TransPort
 echo "_trans_port=\"9040\"" >> $root/iptables_rules.sh
+# Other IANA reserved blocks (These are not processed by tor and dropped by default)
+echo "_resv_iana=\"0.0.0.0/8 100.64.0.0/10 169.254.0.0/16 192.0.0.0/24 192.0.2.0/24 192.88.99.0/24 198.18.0.0/15 198.51.100.0/24 203.0.113.0/24 224.0.0.0/4 240.0.0.0/4 255.255.255.255/32\"" >> $root/iptables_rules.sh
 echo "iptables -F" >> $root/iptables_rules.sh
 echo "iptables -t nat -F" >> $root/iptables_rules.sh
 echo "iptables -t nat -A OUTPUT -d \$_virt_addr -p tcp -m tcp --tcp-flags FIN,SYN,RST,ACK SYN -j REDIRECT --to-ports \$_trans_port" >> $root/iptables_rules.sh
@@ -481,12 +502,19 @@ echo "for _clearnet in \$_non_tor; do" >> $root/iptables_rules.sh
 echo "iptables -t nat -A OUTPUT -d \$_clearnet -j RETURN" >> $root/iptables_rules.sh
 echo "done" >> $root/iptables_rules.sh
 echo "sleep 5" >> $root/iptables_rules.sh
+echo "## Edit and uncomment the next line to grant yourself ssh access from remote machines before the DROP." >> $root/iptables_rules.sh
+echo "#iptables -A INPUT -i YOUR_NETWORK_DEVICE_HERE -p tcp --dport 22 -m state --state NEW -j ACCEPT" >> $root/iptables_rules.sh
 echo "iptables -A INPUT -m state --state ESTABLISHED -j ACCEPT" >> $root/iptables_rules.sh
 echo "iptables -A INPUT -i lo -j ACCEPT" >> $root/iptables_rules.sh
-echo "for _lan in \$_non_tor; do" >> $root/iptables_rules.sh
-echo "iptables -A INPUT -s \$_lan -j ACCEPT" >> $root/iptables_rules.sh
+echo "## Uncomment the next 4 lines to Allow INPUT from lan hosts" >> $root/iptables_rules.sh
+echo "#for _lan in \$_non_tor; do" >> $root/iptables_rules.sh
+echo "#iptables -A INPUT -s \$_lan -j ACCEPT" >> $root/iptables_rules.sh
+echo "#done" >> $root/iptables_rules.sh
+echo "#sleep 5" >> $root/iptables_rules.sh
+echo "for _iana in \$_resv_iana; do" >> $root/iptables_rules.sh
+echo "iptables -t nat -A OUTPUT -d \$_iana -j RETURN" >> $root/iptables_rules.sh
 echo "done" >> $root/iptables_rules.sh
-echo "sleep 5" >> $root/iptables_rules.sh
+echo "sleep 3" >> $root/iptables_rules.sh
 echo "iptables -A INPUT -j DROP" >> $root/iptables_rules.sh
 echo "iptables -A FORWARD -j DROP" >> $root/iptables_rules.sh
 echo "for _clearnet in \$_non_tor; do" >> $root/iptables_rules.sh
@@ -501,6 +529,9 @@ echo "sleep 1" >> $root/iptables_rules.sh
 echo "iptables -P FORWARD DROP" >> $root/iptables_rules.sh
 echo "iptables -P INPUT DROP" >> $root/iptables_rules.sh
 echo "iptables -P OUTPUT DROP" >> $root/iptables_rules.sh
+echo "ip6tables -P INPUT DROP"  >> $root/iptables_rules.sh
+echo "ip6tables -P FORWARD DROP" >> $root/iptables_rules.sh
+echo "ip6tables -P OUTPUT DROP" >> $root/iptables_rules.sh
 chmod +x $root/iptables_rules.sh
 cd $(cat $root/cpath)
 }
@@ -668,20 +699,24 @@ fi
 fi
 fi
 killall leafpad > /dev/null 2>&1
-clear
+echo "";
 echo "==> Configuring DNSCrypt";
+sleep 1
 sed -i "1iforce_tcp = true" $root/dnscrypt-proxy.toml
 sed -i "2iserver_names = ['$server1', '$server2']" $root/dnscrypt-proxy.toml
+sed -i "3iproxy = 'socks5://127.0.0.1:9050'" $root/dnscrypt-proxy.toml
 sed -i "s/127.0.0.1:53/127.0.0.1:10000/g; s/9.9.9.9/208.67.222.222/g; s/8.8.8.8/208.67.220.220/g; s/require_dnssec = false/require_dnssec = true/g; s/force_tcp = false/#force_tcp = false/g; s/\[anonymized_dns\]/\[anonymized_dns\]\nroutes = \[\n{ server_name='$server1', via=\[\'$relay1\', \'$relay2\'\] },\n{ server_name=\'$server2\', via=[\'$relay3\', \'$relay4\'] }\n\]/g; s/skip_incompatible = false/skip_incompatible = true/g" $root/dnscrypt-proxy.toml
-echo "==> Configuring Unbound";
 ### Configuring unbound
+echo "==> Configuring Unbound";
+sleep 1
 unbound-anchor > /dev/null 2>&1
 sleep 1
 echo "server:" > $unbound
 echo "tcp-upstream: yes" >> $unbound
 echo "domain-insecure: \"onion\"" >> $unbound
 echo "private-domain: \"onion\"" >> $unbound
-echo "do-not-query-localhost: no" >> $unbound 
+echo "do-not-query-localhost: no" >> $unbound
+echo "do-ip6: no " >> $unbound
 echo "interface: 127.0.0.1@53" >> $unbound
 echo "local-zone: \"onion.\" transparent" >> $unbound
 echo "forward-zone:" >> $unbound
@@ -690,6 +725,7 @@ echo "    forward-addr: 127.0.0.1@5353" >> $unbound
 echo "forward-zone:" >> $unbound
 echo "   name: \".\"" >> $unbound
 echo "   forward-addr: 127.0.0.1@10000" >> $unbound
+sleep 1
 }
 ##
 ## Starting services and configuring iptables
@@ -705,11 +741,25 @@ if [ -e "menu" ]; then
 menu
 return 1
 else 
+echo "";
 exit 1
 fi
 fi
 if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
+fi
+if [ -s "/etc/network/if-up.d/anon-service" ]; then
+echo "";
+echo "==> Sorry! This menu option is not usable in permanent mode";
+echo "==> Reboot your system or simply restart your connection instead";
+sleep 7
+if [ -e "menu" ]; then
+menu
+return 1
+else 
+echo "";
+exit 1
+fi
 fi
 rm $root/running > /dev/null 2>&1
 ### Firewall flush
@@ -722,39 +772,60 @@ iptables --table nat --delete-chain
 iptables -P OUTPUT ACCEPT
 iptables -P INPUT ACCEPT
 iptables -P FORWARD ACCEPT
-### Disable ipv6 
-echo "1" > /dev/null 2>&1 | tee /proc/sys/net/ipv6/conf/default/disable_ipv6
-echo "1" > /dev/null 2>&1 | tee /proc/sys/net/ipv6/conf/all/disable_ipv6
+ip6tables -P OUTPUT ACCEPT
+ip6tables -P INPUT ACCEPT
+ip6tables -P FORWARD ACCEPT
 ### Configure Network-Manager
 cd $root
-cp resolved.conf.temp $resolved > /dev/null 2>&1
-chown root:root $resolved > /dev/null 2>&1
+if [ -s $netman ]; then
 cp NetworkManager.conf $netman
 chown root:root $netman
+fi
+service resolvconf stop > /dev/null 2>&1
 service dnsmasq stop > /dev/null 2>&1
 service bind stop > /dev/null 2>&1
-service systemd-resolved stop
+service systemd-resolved stop > /dev/null 2>&1
 killall dnsmasq bind > /dev/null 2>&1
 sleep 1
 service tor stop > /dev/null 2>&1
 service dnscrypt-proxy stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
 killall unbound tor dnscrypt-proxy > /dev/null 2>&1
+echo "==> Forcing nameserver to 127.0.0.1 in resolv.conf";
 rm /etc/resolv.conf > /dev/null 2>&1
-service systemd-resolved restart
-service network-manager restart
-sleep 5
+echo $'inameserver 127.0.0.1\E:x\n' | vi /etc/resolv.conf > /dev/null 2>&1
+chattr +i /etc/resolv.conf > /dev/null 2>&1
+sleep 1
 cat /etc/resolv.conf | sed -e '/^$/d; /^#/d' > $root/dnsread
 if [[ $(cat $root/dnsread) != "nameserver 127.0.0.1" ]]; then 
-rm /etc/resolv.conf > /dev/null 2>&1 
 echo "";
-echo "==> Make sure 127.0.0.1 is your DNS system setting and then press ENTER";
-read REPLY
-service systemd-resolved restart
-service network-manager restart
+echo "==> There is a problem with your DNS setting! Fix your";
+echo "==> /etc/resolv.conf by setting 127.0.0.1 as nameserver and try again"
+if [ -e "menu" ]; then
+menu
+return 1
+else 
+echo "";
+exit 1
+fi
 fi
 rm $root/dnsread > /dev/null 2>&1
+echo "==> Restarting networking";
+if [ -f $netman ]; then
+service network-manager restart > /dev/null 2>&1
 sleep 5
+else
+ifdown -a > /dev/null 2>&1 && ifup -a > /dev/null 2>&1
+fi
+sleep 5
+### Disable ipv6 
+ipv6_status=$(cat /proc/sys/net/ipv6/conf/default/disable_ipv6)
+if [ "$ipv6_status" == "0" ]; then
+echo "==> Disabling IPV6 protocol";
+sysctl -w net.ipv6.conf.all.disable_ipv6=1 > /dev/null 2>&1
+sysctl -w net.ipv6.conf.default.disable_ipv6=1 > /dev/null 2>&1
+fi
+sleep 2
 chown -R $owner:$owner $root
 ## Restore original files automatically at shutdown
 if ( ! pgrep -f "restoring_orig.sh " )  > /dev/null; then
@@ -763,8 +834,8 @@ touch restoring_orig.sh
 echo "#!/bin/bash" > restoring_orig.sh
 echo "restoring_script() {" >> restoring_orig.sh 
 echo "if [ ! -f /etc/network/if-up.d/anon-service ]; then" >> restoring_orig.sh
-echo "cp $root/resolved.bak $resolved" >> restoring_orig.sh
-echo "cp $netman.bak $netman" >> restoring_orig.sh
+echo "cp $netman.bak $netman > /dev/null 2>&1" >> restoring_orig.sh
+echo "chattr -i /etc/resolv.conf > /dev/null 2>&1" >> restoring_orig.sh
 echo "fi" >> restoring_orig.sh
 echo "rm $root/running > /dev/null 2>&1" >> restoring_orig.sh
 echo "exit" >> restoring_orig.sh
@@ -775,19 +846,23 @@ echo "trap restoring_script SIGINT SIGTERM" >> restoring_orig.sh
 echo "sleep 7" >> restoring_orig.sh
 echo "done" >> restoring_orig.sh
 chmod +x restoring_orig.sh
-if [ -e "$(cat $root/cpath)/temp/" ]; then
+if [ ! -e "$(cat $root/cpath)/temp/menu" ]; then
 nohup ./restoring_orig.sh > /dev/null 2>&1 &
-clear
+echo "==> Automatic restore started"
+sleep 1
 else
 xterm -e nohup ./restoring_orig.sh
+echo "==> Automatic restore started"
+sleep 1
 fi
 fi
 echo "==> Starting anon-service";
+sleep 1
 ## Start selected transparent proxy
 active_service=$(cat $root/stp-service)
 case $active_service in
 "0")
-nohup su - $owner -c "./dnscrypt-proxy" > /dev/null 2>&1 &
+nohup ./dnscrypt-proxy > /dev/null 2>&1 &
 sleep 1
 rm $root/notices.log > /dev/null 2>&1
 touch $root/notices.log
@@ -806,42 +881,29 @@ done
 rm $root/notices.log > /dev/null 2>&1
 cd $root
 ./iptables_rules.sh
+sleep 2
 unbound &
 ### Checking services
 if ( ! pgrep -x "tor" ) > /dev/null; then
 echo "==> Sorry! No connection to TOR...Please, report this issue to the project";
 sleep 7
-shutdown_service
-if [ -e "$(cat $root/cpath)/temp/menu" ]; then
-menu
-return 1
-else 
+shutdown_service 
 exit 1
-fi
 fi
 if ( ! pgrep -x "dnscrypt-proxy" ) > /dev/null; then
 echo "==> Sorry! Dnscrypt-proxy isn't running...Please, report this issue to the project";
 sleep 7
 shutdown_service
-if [ -e "$(cat $root/cpath)/temp/menu" ]; then
-menu
-return 1
-else
 exit 1
-fi
 fi
 if ( ! pgrep -x "unbound" ) > /dev/null; then
 echo "==> Sorry! Unbound isn't running...Please, report this issue to the project";
 sleep 7
 shutdown_service
-if [ -e "$(cat $root/cpath)/temp/menu" ]; then
-menu
-return 1
-else
 exit 1
-fi
 else
-echo "==> Congratulations! Your system is configurated to use Tor and DNSCrypt";
+echo "==> Service started using Tor and DNSCrypt";
+echo "";
 touch $root/running
 sleep 5
 fi
@@ -867,6 +929,7 @@ cd $root
 ### Checking services
 if ( ! pgrep -x "tor" > /dev/null ); then
 echo "==> Sorry! No connection to TOR...Please, report this issue to the project";
+echo "";
 sleep 7
 shutdown_service
 if [ -e "$(cat $root/cpath)/temp/menu" ]; then
@@ -876,12 +939,89 @@ else
 exit 1
 fi
 else
-echo "==> Congratulations! Your system is configurated to use Tor";
+echo "==> Service started using Tor network";
+echo "";
 touch $root/running
 sleep 5
 fi
 esac
 cd $(cat $root/cpath)
+}
+##
+## Edit configuration files
+##
+editor(){
+echo "   #######################################################";
+echo "   #              ANON-SERVICE CUSTOMIZATION             #";
+echo "   #######################################################";
+echo "";
+echo "==> What do you want to edit?";
+echo "      1.torrc";
+echo "      2.iptables rules";
+echo " "
+echo -n  " Choose: ";
+read -e answer
+echo "";
+case "$answer" in 
+1)
+if [ ! -e "$(cat $root/cpath)/temp/menu" ]; then
+nano $root/torrc
+echo "==> Please restart the service to apply changes";
+exit 0
+else
+xterm -T "Torrc" -e "leafpad $root/torrc" > /dev/null 2>&1 
+echo "==> Please restart the service to apply changes";
+sleep 7
+menu
+fi
+;;
+2)
+if [ -s "/etc/network/if-up.d/anon-service" ]; then
+if [ ! -e "$(cat $root/cpath)/temp/menu" ]; then
+nano /etc/network/if-up.d/anon-service
+echo "==> Please restart via command-line option to apply changes"; 
+echo "==> Otherwise restart your network connection or reboot your system.";
+exit 0
+else
+xterm -T "Editor" -e "leafpad /etc/network/if-up.d/anon-service" > /dev/null 2>&1
+echo "==> Please restart via command-line option to apply changes"; 
+echo "==> Otherwise restart your network connection or reboot your system.";
+sleep 7
+if [ -e "$(cat $root/cpath)/temp/menu" ]; then
+menu
+else 
+exit 0
+fi
+fi
+else 
+if [ ! -e "$(cat $root/cpath)/temp/menu" ]; then
+nano $root/iptables_rules.sh
+echo "==> Please restart the service to apply changes";
+exit 0
+else
+xterm -T "Editor" -e "leafpad $root/iptables_rules.sh" > /dev/null 2>&1
+echo "==> Please restart the service to apply changes";
+sleep 7
+if [ -e "$(cat $root/cpath)/temp/menu" ]; then
+menu
+else 
+exit 0
+fi
+fi
+fi
+;;
+*)
+echo "==> Are you serious?"
+sleep 5
+if [ -e "$(cat $root/cpath)/temp/menu" ]; then
+menu
+return 1
+else 
+echo "";
+exit 1
+fi
+;;
+esac
 }
 ##
 ## Install this script
@@ -892,7 +1032,9 @@ cp $0 /usr/bin/anon-service > /dev/null 2>&1
 chmod +x /usr/bin/anon-service
 echo "";
 echo "==> Now you can run it simply typing \"sudo anon-service\" in your terminal";
+echo "";
 cd $(cat $root/cpath)
+exit 0
 }
 ##
 ## Run at boot
@@ -907,16 +1049,19 @@ if [ -e "$(cat $root/cpath)/temp/menu" ]; then
 menu
 return 1
 else
+echo "";
 exit 1
 fi
 fi
 cd $root
-cp resolved.conf.temp $resolved > /dev/null 2>&1
-chown root:root $resolved > /dev/null 2>&1
-cp NetworkManager.conf $netman
-chown root:root $netman
+cp NetworkManager.conf $netman > /dev/null 2>&1
+chown root:root $netman > /dev/null 2>&1
 rm /etc/network/if-up.d/anon-service > /dev/null 2>&1
 touch /etc/network/if-up.d/anon-service
+chattr -i /etc/resolv.conf > /dev/null 2>&1
+rm /etc/resolv.conf > /dev/null 2>&1
+echo $'inameserver 127.0.0.1\E:x\n' | vi /etc/resolv.conf > /dev/null 2>&1
+chattr +i /etc/resolv.conf > /dev/null 2>&1
 echo "#!/bin/sh" > /etc/network/if-up.d/anon-service
 echo "root=/home/anon-service" >> /etc/network/if-up.d/anon-service
 echo "owner=anon-service" >> /etc/network/if-up.d/anon-service
@@ -929,11 +1074,14 @@ echo "iptables --delete-chain" >> /etc/network/if-up.d/anon-service
 echo "iptables --table nat --delete-chain" >> /etc/network/if-up.d/anon-service 
 echo "iptables -P OUTPUT ACCEPT" >> /etc/network/if-up.d/anon-service
 echo "iptables -P INPUT ACCEPT" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P FORWARD ACCEPT" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P OUTPUT ACCEPT" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P INPUT ACCEPT" >> /etc/network/if-up.d/anon-service
 echo "iptables -P FORWARD ACCEPT" >> /etc/network/if-up.d/anon-service
-echo "echo \"1\" > /dev/null 2>&1 | tee /proc/sys/net/ipv6/conf/default/disable_ipv6" >> /etc/network/if-up.d/anon-service
-echo "echo \"1\" > /dev/null 2>&1 | tee /proc/sys/net/ipv6/conf/all/disable_ipv6" >> /etc/network/if-up.d/anon-service
 echo "service dnsmasq stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "service bind stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
+echo "service resolvconf stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
+echo "service systemd-resolved stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "killall dnsmasq bind > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "sleep 1s" >> /etc/network/if-up.d/anon-service
 echo "cd $root" >> /etc/network/if-up.d/anon-service
@@ -941,12 +1089,9 @@ echo "service tor stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "service dnscrypt-proxy stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "service unbound stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "killall unbound tor dnscrypt-proxy > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
+echo "sysctl -w net.ipv6.conf.all.disable_ipv6=1 > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
+echo "sysctl -w net.ipv6.conf.default.disable_ipv6=1 > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "chown -R $owner:$owner $root" >> /etc/network/if-up.d/anon-service
-selected_service=$(cat $root/stp-service)
-if (( $selected_service == 1 )); then
-echo "nohup su - $owner -c \"./dnscrypt-proxy\" > /dev/null 2>&1 &" >> /etc/network/if-up.d/anon-service
-echo "sleep 1s" >> /etc/network/if-up.d/anon-service
-fi
 echo "rm $root/notices.log > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "touch $root/notices.log" >> /etc/network/if-up.d/anon-service
 echo "chown anon-service:anon-service $root/notices.log" >> /etc/network/if-up.d/anon-service
@@ -959,11 +1104,17 @@ echo "else" >> /etc/network/if-up.d/anon-service
 echo "sleep 1s" >> /etc/network/if-up.d/anon-service
 echo "fi" >> /etc/network/if-up.d/anon-service
 echo "done" >> /etc/network/if-up.d/anon-service
+selected_service=$(cat $root/stp-service)
+if (( $selected_service == 0 )); then
+echo "nohup ./dnscrypt-proxy > /dev/null 2>&1 &" >> /etc/network/if-up.d/anon-service
+echo "sleep 1s" >> /etc/network/if-up.d/anon-service
+fi
 echo "rm $root/notices.log > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "_non_tor=\"127.0.0.0/8 10.0.0.0/8 172.16.0.0/12 192.168.0.0/16\"" >> /etc/network/if-up.d/anon-service
 echo "_user_uid=\"999\"" >> /etc/network/if-up.d/anon-service
 echo "_virt_addr=\"10.192.0.0/10\"" >> /etc/network/if-up.d/anon-service
 echo "_trans_port=\"9040\"" >> /etc/network/if-up.d/anon-service
+echo "_resv_iana=\"0.0.0.0/8 100.64.0.0/10 169.254.0.0/16 192.0.0.0/24 192.0.2.0/24 192.88.99.0/24 198.18.0.0/15 198.51.100.0/24 203.0.113.0/24 224.0.0.0/4 240.0.0.0/4 255.255.255.255/32\"" >> /etc/network/if-up.d/anon-service
 echo "iptables -F" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -F" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -A OUTPUT -d \$_virt_addr -p tcp -m tcp --tcp-flags FIN,SYN,RST,ACK SYN -j REDIRECT --to-ports \$_trans_port" >> /etc/network/if-up.d/anon-service
@@ -979,12 +1130,19 @@ echo "for _clearnet in \$_non_tor; do" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -A OUTPUT -d \$_clearnet -j RETURN" >> /etc/network/if-up.d/anon-service
 echo "done" >> /etc/network/if-up.d/anon-service
 echo "sleep 5s" >> /etc/network/if-up.d/anon-service
+echo "## Edit and uncomment the next line to grant yourself ssh access from remote machines before the DROP." >> /etc/network/if-up.d/anon-service
+echo "#iptables -A INPUT -i YOUR_NETWORK_DEVICE_HERE -p tcp --dport 22 -m state --state NEW -j ACCEPT" >> /etc/network/if-up.d/anon-service
 echo "iptables -A INPUT -m state --state ESTABLISHED -j ACCEPT" >> /etc/network/if-up.d/anon-service
 echo "iptables -A INPUT -i lo -j ACCEPT" >> /etc/network/if-up.d/anon-service
-echo "for _lan in \$_non_tor; do" >> /etc/network/if-up.d/anon-service
-echo "iptables -A INPUT -s \$_lan -j ACCEPT" >> /etc/network/if-up.d/anon-service
+echo "## Uncomment the next 4 lines to Allow INPUT from lan hosts" >> /etc/network/if-up.d/anon-service
+echo "#for _lan in \$_non_tor; do" >> /etc/network/if-up.d/anon-service
+echo "#iptables -A INPUT -s \$_lan -j ACCEPT" >> /etc/network/if-up.d/anon-service
+echo "#done" >> /etc/network/if-up.d/anon-service
+echo "#sleep 5s" >> /etc/network/if-up.d/anon-service
+echo "for _iana in \$_resv_iana; do" >> /etc/network/if-up.d/anon-service
+echo "iptables -t nat -A OUTPUT -d \$_iana -j RETURN" >> /etc/network/if-up.d/anon-service
 echo "done" >> /etc/network/if-up.d/anon-service
-echo "sleep 5s" >> /etc/network/if-up.d/anon-service
+echo "sleep 3s" >> /etc/network/if-up.d/anon-service
 echo "iptables -A INPUT -j DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -A FORWARD -j DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -A OUTPUT -p tcp --syn -j REDIRECT --to-ports \$_trans_port" >> /etc/network/if-up.d/anon-service
@@ -999,7 +1157,11 @@ echo "sleep 1s" >> /etc/network/if-up.d/anon-service
 echo "iptables -P FORWARD DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -P INPUT DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -P OUTPUT DROP" >> /etc/network/if-up.d/anon-service
-if (( $selected_service == 1 )); then
+echo "sleep 1s" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P FORWARD DROP" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P INPUT DROP" >> /etc/network/if-up.d/anon-service
+echo "ip6tables -P OUTPUT DROP" >> /etc/network/if-up.d/anon-service
+if (( $selected_service == 0 )); then
 echo "unbound" >> /etc/network/if-up.d/anon-service
 fi
 echo "echo \"+++ anon-service started +++\"" >> /etc/network/if-up.d/anon-service
@@ -1008,8 +1170,8 @@ chown root:root /etc/network/if-up.d/anon-service
 chmod 755 /etc/network/if-up.d/anon-service
 chmod +x /etc/network/if-up.d/anon-service
 echo "";
-echo "==> Now you are ready to go! If you haven't set 127.0.0.1 in your DNS"; 
-echo "==> setting, do it and restart your connection or reboot your system.";
+echo "==> Now you are ready to go! Use the restart command-line option"; 
+echo "==> Otherwise restart your network connection or reboot your system.";
 echo "";
 cd $(cat $root/cpath)
 }
@@ -1017,8 +1179,13 @@ cd $(cat $root/cpath)
 ## CHECKING IF RUNNING
 ##
 checking_service(){
+echo "   #######################################################";
+echo "   #                 ANON-SERVICE STATUS                 #";
+echo "   #######################################################";
+echo "";
 if [ ! -e $root/running ] > /dev/null; then
 echo "==> Service is not running!";
+echo "";
 sleep 3	
 else
 echo "==> Service is running!";
@@ -1026,12 +1193,14 @@ curl --socks5 localhost:9050 --socks5-hostname localhost:9050 -s https://check.t
 if ( grep -q "Your" $root/ip.txt ); then
 ipaddr=$(cat $root/ip.txt)
 echo "==> $ipaddr";
+echo "";
 else
 echo "==> But the service can't access internet. Try the restart option";
 if [ -e "$(cat $root/cpath)/temp/menu" ]; then
 menu
 return 1
 else
+echo "";
 exit 1
 fi
 fi
@@ -1041,10 +1210,14 @@ fi
 ## Exit
 ##
 shutdown_service(){
+banner
+echo "";
 if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
+if [ -e "$(cat $root/cpath)/temp/menu" ]; then
 clear
+fi
 service dnscrypt-proxy stop > /dev/null 2>&1
 sleep 3
 if ! pgrep -x "tor" > /dev/null; then
@@ -1052,26 +1225,34 @@ echo "==> Restoring original files";
 sleep 7
 else
 echo "==> Stopping anon-service";
-sleep 7
+sleep 2
+echo "==> Restoring original files";
 fi
 rm $root/tor.txt > /dev/null 2>&1
 rm $root/running > /dev/null 2>&1
-service dnscrypt-proxy stop > /dev/null 2>&1
+chattr -i /etc/resolv.conf > /dev/null 2>&1
+rm /etc/resolv.conf > /dev/null 2>&1
 service tor stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
-killall unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
-cp $root/resolved.bak $resolved > /dev/null 2>&1
+killall xterm unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
 cp $netman.bak $netman > /dev/null 2>&1
+echo "==> Restarting neworking";
+service systemd-resolved restart
+if [ -f $netman ]; then
+service network-manager restart > /dev/null 2>&1
+sleep 1
+else
+ifdown -a > /dev/null 2>&1 && ifup -a > /dev/null 2>&1
+sleep 1
+fi
 if [ -s "/etc/network/if-up.d/anon-service" ]; then
 rm /etc/network/if-up.d/anon-service
 echo "";
 echo "==> Now the service is no more enabled at startup!";
 echo "==> You can reactivate it using appropriate option";
-echo -e "\n\n";
+echo "";
 sleep 7
 fi
-service systemd-resolved restart
-service network-manager restart
 ### Firewall flush
 iptables -F
 iptables -t nat -F
@@ -1082,28 +1263,30 @@ iptables --table nat --delete-chain
 iptables -P OUTPUT ACCEPT
 iptables -P INPUT ACCEPT
 iptables -P FORWARD ACCEPT
+ip6tables -P OUTPUT ACCEPT
+ip6tables -P INPUT ACCEPT
+ip6tables -P FORWARD ACCEPT
 cd $(cat $root/cpath)
 }
 ##
 ## Cleaning all and exit
 ##
 cleanall(){
+if [ -e "$(cat $root/cpath)/temp/menu" ]; then
 clear
-echo "==> Removing anon-service files and settings from system";
+fi
+echo "   #######################################################";
+echo "   #               ANON-SERVICE UNINSTALLER              #";
+echo "   #######################################################";
+echo "";
+echo "==> Stopping service";
+sleep 1
 service tor stop > /dev/null 2>&1
-service dnscrypt-proxy stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
-killall unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
-if [[ -s "$root/resolved.bak" ]]; then
-cp $root/resolved.bak $resolved > /dev/null 2>&1
-service systemd-resolved restart
-fi
-if [[ -s "$netman.bak" ]]; then
-cp $netman.bak $netman > /dev/null 2>&1
-fi
-rm /usr/bin/anon-service > /dev/null 2>&1
-service systemd-resolved restart
-service network-manager restart
+killall xterm unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
+echo "==> Removing anon-service files and settings from system";
+chattr -i /etc/resolv.conf > /dev/null 2>&1
+rm /etc/resolv.conf > /dev/null 2>&1
 rm $repo > /dev/null 2>&1
 rm $repo* > /dev/null 2>&1
 rm /etc/network/if-up.d/anon-service > /dev/null 2>&1
@@ -1119,6 +1302,19 @@ rm -rf $(cat $root/cpath)/temp > /dev/null 2>&1
 userdel -r $owner > /dev/null 2>&1
 rm -rf $root > /dev/null 2>&1
 rm cpath > /dev/null 2>&1
+echo "==> Restarting neworking";
+service systemd-resolved restart
+if [[ -s "$netman.bak" ]]; then
+cp $netman.bak $netman > /dev/null 2>&1
+fi
+rm /usr/bin/anon-service > /dev/null 2>&1
+if [ -f $netman ]; then
+service network-manager restart > /dev/null 2>&1
+sleep 1
+else
+ifdown -a > /dev/null 2>&1 && ifup -a > /dev/null 2>&1
+sleep 1
+fi
 ### Firewall flush
 iptables -F
 iptables -t nat -F
@@ -1129,9 +1325,11 @@ iptables --table nat --delete-chain
 iptables -P OUTPUT ACCEPT
 iptables -P INPUT ACCEPT
 iptables -P FORWARD ACCEPT
+ip6tables -P OUTPUT ACCEPT
+ip6tables -P INPUT ACCEPT
+ip6tables -P FORWARD ACCEPT
 clear
 echo -e "\n\n\n\n\n\n";
-echo "==> Remember to change your DNS system setting";
 echo "    ______________________________________________";
 echo " "
 echo "    +++ Have a nice day! ;) +++";
@@ -1156,18 +1354,37 @@ printf '%s\n' "                      -2 OS repository -3 already installed"
 printf '%s\n' " --configure <value>  choose transparent proxy type"
 printf '%s\n' "                      <value> -1 standard -2 with DNSCrypt"
 printf '%s\n' " --start              start service"
-printf '%s\n' " --stop               without removing service files and settings"
+printf '%s\n' " --stop               exit without removing service files and settings"
 printf '%s\n' " --restart            restart service"
 printf '%s\n' " --status             display status service"
 printf '%s\n' " --menu               display interactive menu"
 printf '%s\n' " --install            install this script"
 printf '%s\n' " --permanent          enable service to start automatically at boot"
 printf '%s\n' " --remove             exit removing files and settings from system"
-printf '%s\n' " --edit               edit torrc file"
+printf '%s\n' " --edit      <value>  edit configuraion files"
+printf '%s\n' "                      <value> torrc or iptables"
+printf '%s\n' " --restore            restore original files and settings"
 echo "";
 printf '%s\n' " --help               display this help"
 printf '%s\n' " --version            display version"
 echo "";
+}
+##
+## Banner
+##
+banner(){
+printf '%s\n' "           ▄▄▄      ███▄    █ ▒█████   ███▄    █          "
+printf '%s\n' "          ▒████▄    ██ ▀█   █▒██▒  ██▒ ██ ▀█   █          "
+printf '%s\n' "          ▒██  ▀█▄ ▓██  ▀█ ██▒██░  ██▒▓██  ▀█ ██▒         "
+printf '%s\n' "          ░██▄▄▄▄██▓██▒  ▐▌██▒██   ██░▓██▒  ▐▌██▒         "
+printf '%s\n' "           ▓█   ▓██▒██░   ▓██░ ████▓▒░▒██░   ▓██░         "
+printf '%s\n' "       ██████ ▓█████  ██▀███░  ██▒ ░ █▓ ██▓ ▄████▄ ▓█████ "
+printf '%s\n' "     ▒██    ▒ ▓█   ▀ ▓██ ▒ ██▒▓██░   █▒▓██▒▒██▀ ▀█ ▓█   ▀ "
+printf '%s\n' "     ░ ▓██▄   ▒███   ▓██ ░▄█ ▒ ▓██  █▒░▒██▒▒▓█    ▄▒███   "
+printf '%s\n' "       ▒   ██▒▒▓█  ▄ ▒██▀▀█▄    ▒██ █░░░██░▒▓▓▄ ▄██▒▓█  ▄ "
+printf '%s\n' "     ▒██████▒▒░▒████▒░██▓ ▒██▒   ▒▀█░  ░██░▒ ▓███▀ ░▒████▒"
+printf '%s\n' "           ░           ░           ░       ░ by bit4mind  "
+echo " ";
 }
 ##
 ## Main
@@ -1307,6 +1524,7 @@ configure
 fi
 ;;
 --start)
+banner
 if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
@@ -1321,30 +1539,34 @@ shutdown_service
 exit 0
 ;;
 --restart)
-echo "==> Reloading...";
-sleep 3
+echo "Reloading...";
+sleep 1
+clear
+banner
 if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
+if [ -s "/etc/network/if-up.d/anon-service" ]; then
+if [ -f $netman ]; then
+service network-manager restart > /dev/null 2>&1
+echo "";
+echo "==> Service restarted...Be patient!";
+echo "";
+else
+ifdown -a > /dev/null 2>&1 && ifup -a > /dev/null 2>&1
+echo "";
+echo "==> Service restarted...Be patient!";
+echo "";
+fi
+exit 0
+else
 start_service
 exit 0
+fi
 ;;
 --status)
-if [ ! -e $root/running ] > /dev/null; then
-echo "==> Service is not running!";
-sleep 3	
-else
-echo "==> Service is running!";
-curl --socks5 localhost:9050 --socks5-hostname localhost:9050 -s https://check.torproject.org/ | cat | grep -m 	1 "Your IP address" | sed -e 's/<[^>]*>//g' | xargs > $root/ip.txt
-sleep 3
-if ( grep -q "Your" $root/ip.txt ); then
-		ipaddr=$(cat $root/ip.txt)
-echo "==> $ipaddr";
-else
-echo "==> But the service can't access internet. Try the restart option";
+checking_service
 exit 0
-fi
-fi
 ;;
 --menu)
 cd temp
@@ -1362,7 +1584,12 @@ exit 0
 cleanall
 exit 0
 ;;
+--restore)
+shutdown_service
+exit 0
+;;
 --version)
+banner
 echo "anon-service $version";
 ;;
 --help)
@@ -1379,11 +1606,33 @@ echo "==> Sorry! Your system is not ready to complete this action";
 echo "==> Please, check if you have installed the necessary files";
 sleep 3
 exit 1
-else
-echo "";
+fi
+if [ ! -z "$2" ]; then
+case "$2" in
+torrc)
 nano $root/torrc
 echo "==> Please restart the service to apply changes";
-sleep 3
+exit 0
+;;
+iptables)
+if [ -s "/etc/network/if-up.d/anon-service" ]; then
+nano /etc/network/if-up.d/anon-service
+echo "==> Please restart via command-line option to apply changes"; 
+echo "==> Otherwise restart your network connection or reboot your system.";
+exit 0
+else 
+nano $root/iptables_rules.sh
+echo "==> Please restart the service to apply changes";
+exit 0
+fi
+;;
+*)
+echo "==> Invalid option '$2'";
+exit 1
+;;
+esac
+else
+editor
 exit 0
 fi
 ;;